Rewrite ExtractExpressions
We need to have a proper scope to actually extract primitive
expressions. Otherwise we don't know if symbols are refering to the
(scheme core) library or something else. At the same time add some
provisional hand-wavy support for scopes, libraries and programs

Lamdba support is disabled as it requires proper body parsing.
(define-syntax) and (define) support are removed because they aren't
primitive expressions