Make ExtractBody a loop instead of recursive
ExtractBody was never actually tail recursive which means it would
eventually explode with large bodies. The passing of restData was also a
bit cumbersome. Switch to making apply() a wrapper that does a foldLeft
over the data. Using a ListBuffer is a bit dirty; this could be done
with a List reversed at the end but the code was harder to read.