Make the AST more consistent
Rename Vector to VectorLiteral to prevent shadowing
scala.collection.immutable.Vector

ProperList was applying varargs but unapplying a List. That means that
we were extracting something different than we were constructing which
breaks the pattern matching metaphor

The ProperList extractor syntax is too powerful to give up by switching
it to unapplySeq. Instead switch all the AST types to take a sequence
instead of using varargs. Make ByteVector/VectorLiteral use a Scala
Vector both for analogy with the Scala types and possible usage patterns
if we deeply inspect them later in compilation