Change nfi.Bool to nfi.Bool32
nfi.Bool was previously defined as i1. Predicates are expressed in LLVM
IR as i8 but the x86-64 ABI packs it in an i32. We should be using the
i32 version when describing function signatures.

At the same time add an alias from "bool" => nfi.Bool32 in case we port
to an ABI that uses a different sized bool