Add IrNameSource
This handles the allocation of local and global variable names.
Previously IrBlock was allocating them via IrBuilder. Unfortunatey,
local variables have function scope. This means in a function with
multiple blocks the variables would immediately conflict

It also adds support for using a basename if a more descriptive variable
name is wanted