Add JSON description of our boxed runtime types
This describes the type ID, structure and inheritance tree of our
boxed types. This is intended as the source for generated code in the
compiler, runtime and LLVM prelude for dealing with types. Actual code
generator will follow in later commits.

Special syntax is used for the vectorLike type ID: "&32768". This means
that any type ID with its high bit set is considered a vectorLike.
vectorLike is abstract so it's impossible to directly instantiate it.
Instead, one of its subtypes with explicit type IDs must be used.

The field type system is based on LLVM IR types except for two cases:
- Procedure entry points have elaborate types (i.e. function pointers)
  that would be easier for each code generator to special case
- Booleans can only have a value of "0" or "1".  Knowing this allows
  LLVM to optimize their use more than a bare i8. We could use range
  annotations like LLVM IR does but that wouldn't translate in to
  generated C++