Fix Scheme serialization of AbstractSyntaxTree
This extends the SchemeParserSuite to use a new method called
assertReflexiveParse. First, it parses the passed string and verifies it
matches the expected value. Then it serializes the parsed data back to a
Scheme string and re-parses it. Finally it confirms the re-parsed data
still matches the expected value.

assertReflexiveParse caught bugs in the serialization of reals, symbols
and characters. This fixes all found issues.