Runtime support for unspecific, bool, empty list
This follows the Llispy model of using singletons for boolean true and
false and extends it to also cover unspecific and empty list. This has
the original benefit of being able to test truthiness with a pointer
compare but should additionally reduce garbage collector load as
unspecific and empty list are common values.

To reduce code duplication in the binding a new SingletonValue class is
created that passes the GlobalConstant GC state to BoxedDatum and
deletes the new operator.

New core/constinstances.{h,cpp} are created to actually host the
singleton instances. These are also available from static methods on
their respective binding classes.

For now externally_available along with the inline definition is removed
from defines.ll. The boxed types are still in flux so hardcoding their
initialized state and type IDs seems like premature optimization.
Additionally I can't think of a case where we'd generate code that would
cause LLVM to use the inline definition.