Make compiler pass the correct stdlib to clang++
Our runtime requires C++11 features that only libc++ supports on Mac OS
X due to the ancient version of stdlibc++ shipped on that platform. The
CMakeLists.txt for the runtime already passes -stdlib=libc++ on Mac OS X
but it also needs to be passed to the clang++ our compiler invokes or
the wrong library will be linked.