Add lliby_init
lliby_init is to be called before calling any other runtime functions.
Currently it just calls setlocale() but in the future it may initialize
the garbage collector etc.

This can be done with global constructors without an explicit entry
point but the ordering is messy and fragile. An explicit init function
seems cleaner.