Add asciiOnlyHint to StringLike
Many operations such as length in characters, accessing individual
characters, etc. are O(1) on ASCII strings but O(n) on UTF-8 strings.
The asciiOnlyHint can be set by the compiler when it constructs string
it knows to be ASCII only. The runtime can then optionally use that to
accelerate certain operations.

There may be situations where it's hard or computationally difficult to
determine if a string contains ASCII characters. That means it's
possible that asciiOnlyHint is false on a string that in fact contains
ASCII characters.