Add simple CTest-based test support
This directly uses CTest along with a set of simple assertions in
assertions.h. This is intended to test lower-level functionality such as
StringLike functionality and garbage collector internals that are hard
to directly test from Scheme. The stdlib will generally be tested from
compiled Scheme programs created through ScalaTest.