Add alloc::FreezeGc
This freezes the garbage collector with the specified number of conses
pre-allocated. This is intended to be used when building complex
structures incrementally where a GC pass might erroneously free the
partial structure before it can be returned and become reachable.

Due to lack of a GC implementation this is currently a stub.