Build BoxedType methods to generate IR type checks
This extends gen-types.py to build type checks for each boxed type
except BoxedDatum. For most types this simply loads the type ID pointer
and does a direct integer comparison. However, for supertypes it will
either check the type ID of each of its subtypes to use a bitmask if one
is defined

At the same time this renames the "pointerTo*" accessors to
"genPointerTo*" to denote that they generate code.