Replace et.Let with et.Bind
et.Let mirrored the behavior of R7RS's  (let) where it associated
bindings with a list of expressions over which they were effective. This
caused some difficulty with loading libraries as loaded libraries have
their bindings effective past the end of the library

Instead, introduce et.Bind which does not confine bindings to a
particular region. The bindings are understood to be valid until the end
of the lambda.

(let) is still implemented by the frontend in a way that preserves the
R7RS sematics. This only changes the expression tree generated by the
frontend