Compile expressions from libraries
The expressions collected by the library loader were exposed from
DefaultLibraryLoader.libraryExpressions but they were not used elsewhere
in the compiler.

Instead of passing around a simple function to load libraries add a
LibraryLoader abstract class which defines a load function and a
libraryExpressions function. At the end of ExtractProgram prefix the
libraryExpressions to the body expressions when returning the expression
list.