Support for unboxing to float/double
This is tricky as both exact integers and inexact rationals can be
unboxed to floating point types. If the exact type is known we will
unboxed directly. However, if the type is numeric or one of its
superclasses we generate branches for both ints and rationals and then
phi the result of the unboxing together at the end.