Make BoxedLiveValue use arbitrary possible types
Previously BoxedLiveValue took a specific boxed type and derived a list
of possible concrete boxed types from that specific type. Instead have
NativeToLiveValue calculate the list of possible types and pass that in
to BoxedLiveValue. In the future this will allow BoxedLiveValue to have
arbitrary sets of boxed types. An example of this is conditionals where
the boxed value can have the union of possible types from the true and
false branches.