Support for boxing and unboxing chars and strings
This required some reworking to always pass the module to
ConstantLiveValue instances. This is needed because strings need to
generate a global variable even for unboxed values. Either the module
needs to be passed in to the genUnboxedConstant partial function which
complicates the signature or the module can be passed in before. The
latter seemed cleaner