Add FlattenSelfExecutingLambdas optimizer pass
This turns trivial self-executing lambdas in to et.Begin expressions
that assign the functions parameters to variables and then run the
former lambda's body.

This is required at this early stage because many macros such as (begin)
and (let) are defined in terms of self-executing lambdas. Before this
commit they would cause code generation failures due to unimplemented
lambda support. In the future this optimization pass will still be
beneficial to avoid function generation/call overhead and preserve type
information.