codegen support for cons allocation
Add GenConsAllocation which will allocate the specified number of
conses. It will attempt to directly allocate using the exported
_lliby_alloc_{start, end} pointers and if that fails it will fallback to
invoking the runtime's allocator to perform garbage collection.

This result is wrapped in a ConsAllocation which will only return
pointers initialized to a specific type ID. It also verifies that the
requested cons is in range for the allocation size.