Replace FreezeGc with RangeAlloc
RangeAlloc is a wrapper class around alloc::allocateCons() that has
iterators for iterating over the allocated area. This has a few
advantages:

- Like FreezeGc it requires calculating the number of Conses up front.
  However, as long as only one allocation is made at a time a global
  freeze count doesn't need the be maintained. FreezeGc could
  technically be nested but the nested FreezeGc blocks had no way to
  ensure their allocation. It's safer to only allow the one level of
  nesting that can be guaranteed by the GC.

- The pointer arithmetic is done internally in Cons increments.
  (make-list) was using BoxedPair-based arithmetic which happens to
  work but is technically incorrect and fragile if Cons changes.

- In debug builds the range iterator is bounds checked. In release
  builds the code should compile away to simple pointer arithmetic.

In addition this changes BoxedListElement to layout the list in
forward sequential order in memory. CPUs generally have an easier time
performing forward memory prefetches so this should be a slightly more
optimial layout.