Directly box integers, rationals, chars in codegen
While this avoids the function call overhead in to the stdlib, the
greatest benefit comes from giving LLVM more information about memory
accesses. Previously it would have to reload all memory each time a
value was boxed because it had no insight in to what the boxing
functions were doing. Additionally LLVM will now have more information
about the values behind the pointers. This might allow it to elide a
load if the value is unboxed shortly after boxing, for example.

Keep calling to the stdlib to unbox UTF-8 strings. There's a
considerable amount of work to box a UTF-8 string and it should be
fairly rare operation.