Use (define-report-procedure) in (scheme core)
This annotates the bindings we introduce as complying to the behavior
defined in R7RS. This is currently a noop but allows optimizations to be
performed in the future based on the expected semantics of the
procedures.

define-report-procedure is locally renamed to define-r7rs to save some
horizontal space. This has no semantic effect.