Include the source location in our AST, SST
This is used for more informative error messages. We don't currently
pass this past the expression tree due to conflicts with the upcoming
codegen refactor. We also don't track the expansion location inside of
the expaneded macros; I don't have a good idea what the best way to
track and display those locations are yet.