Clean up application extraction
Previously syntax and primitives were implemented by expanding an
et.VarRef from their symbol and then looking inside the et.VarRef in
extractApplication.  The problem is only StorageLocations are truly
variable and can be referenced outside of the fiction of the frontend.
Referencing syntax or primitives as values is nonsensical but et.VarRef
had to allow it to simulate their application.

Clean up by changing extractApplication to extractSymbolApplication
which takes a scoped symbol and looks up its bound value. This can
handle syntax, primitives and storage locations. If there is instead a
nested expression in the procedure position that expression will be
recursively extracted and the et.Apply manually built without help from
extractSymbolApplication.  Because it's not possible to use syntax
or primitives as expressions the nested resolution case does not need to
handle them.

Clean up et.VarRef to only take a StorageLocation which cleans up the
expression tree for non-frontend users

Finally, fix unit tests that were abusing using primitives as
expressions. Add two new tests to verify that syntax and expressions
cannot be used as expressions.