Fix ambigious parse of unrecongized symbolic chars
Alphabetic literals cannot be followed by a character allowed in an
identifier. This is to prevent unrecognized symbolic characters being
treated as their first character followed by a symbol composed of the
rest of their characters. Without this rule "#\lambda" would be treated
as "#\l ambda", for example.

This behavior is described in R7RS section 6.6