Introduce the planner
The planner takes an expression tree and turns it in to a sequential,
single assignment set of plan steps. These plan steps are then passed to
codegen to be converted to LLVM IR. The introduction of the plan steps
allows some mid-level optimizations such as coalescing GC allocations or
removing redundant boxing that's not possible at the expression tree
level or at the LLVM level.