Clean up temporary test executables
I was under the impression that File.createTempFile implied
deleteOnExit. This was clearly not the case. Additionally, we should try
to delete the test exectuable as soon as we're finished using it anyway.

Both call deleteOnExit() and wrap the test execution in a try/finally
that will delete the executuable. This both prevents garbage from
accumulating and likely reduces disk writes due to most of the tests
being deleted before they have a chance to be scheduled for write.