Add ValueNotApplicableException
This converts the InternalCompilerError in the Planner to a full fledged
SemanticException.

Originally this was intended to be caught in the frontend and anything
that reached what was at the time part of codegen would be an ICE.
However, this code has been moved to the planner and only it has enough
type information to determine if something is definitely not applicable.
This is closely analogous to ImpossibleTypeConversionException which is
also caught in the planner.

Making this a LocatedSemanticException also gives us proper source
location reporting for free.