Add (typeless-boxed) to (llambda test-util)
This is a macro that expands to the (car (cons <foo> #f)) pattern to box
a value and erase its type information. In addition to making the intent
of the unit tests more clear this also allows for easy replacement once
our optimizer is clever enough to optimize away tricks like that.