Replace StepBuffer with PlanWriter
StepBuffer was just a ListBuffer of ps.Steps. This was fine if we're
only generating a single function but falls apart quickly otherwise.

PlanWriter contains a "step" member which is analogous to the old
StepBuffer. Additionally, it contains a mutable map which allows any
method that can output steps to define new functions. This is required
because IntermediateValue needs to generate stub trampolines to for
procedure boxing deep inside its implementation.