Support for first-class procedures
This adds runtime boxing and unboxing of procedures. If a procedures
signature doesn't match the generic BoxedProcedureSignature it will have
a trampoline built at runtime to adapt the signature.

Currently closures and user-defined procedures are not supported. Only
NFI functions can be boxed and unboxed. This provides much of the
framework for user-defined procedures, however.