Remove the StrictBool/TruthyBool distinction
"strict-bool" can easily be replaced with "boxed-boolean" with little to
no overhead. There's also a bit of conceptual load for strict versus
truthy which Vonly seems useful for boolean=?

This is needed because types will be converted from context-dependent
keywords to proper bound values. The current struct/truthy distinction
behaves differently depending if the type is a parameter or return
value which doesn't work well with the bound value model.