Add concept of internal types
Internal types can be still referenced by the runtime and codegen but
aren't exposed to the NFI. Examples of internal types are string-like
(which are a runtime implementation detail) and mutable-var (which is
used to take mutable variables to the GC).

As types are about to be placed in the same namespace as variables and
syntax this helps avoid namespace pollution as well as hiding internal
implementation details.