Change native types in to bound values
This changes native types from syntatic keywords to full bound values.
This is done for symmetry with (define-record-type) which also binds the
resulting type to a value. Having both intrinsic types and user-defined
types in the same namespace makes it easy to define compound types
containing both instrinsic and user-defined types.