SplitRecordType to BoxedRecordType, RecordDataType
vt.RecordDataType is the unboxed version of a record. In signatures this
means its either taking or returning a bare pointer to the record data
itself.

vt.BoxedRecordType is the boxed version of a given record. It's
different from vt.BoxedValue(bt.BoxedRecord) in that it's strongly typed
to a specific record type.