IntermediateValue support for record types
This completes the planner support for boxed type. It also removes
support for raw record data types as there's no sane way to make
automatic conversion to/from raw data work with the garbage collector.
The current intent is to add an <unmanaged-pointer> type and add
procedures to explicitly convert records to/from unmanaged pointers