Support TBAA for record types
Assume every record field cannot alias anything except itself. This is
a very strong assumption but we can guarantee it because we don't
allow arbitrary pointer manipulation that would allow a program to
overlap a record with other memory.

This is bordering on microoptimization but will become important if
mutables are re-implemented on top of records.