Support for closing vars by non-captures
In this context "non-capture" means a variable that does not need to be
represented in the closure. Examples of non-captured variables are other
procedures without closures or constants.

The practical application of this is that (scheme base) procedures can
be called from lambdas. This is the beginning of lambdas being useful.