Support for capturing variables
This changes the generated function's closure arg back to a "self" arg.
The caller can't easily discover if the procedure has packed its record
fields inline. It has to pass a pointer to the procedure cell and have
the function itself unpack it.

This also implements half-built procedure cells that only contain a
closure but no entry point. These are represented by iv.KnownProcedure.
This can be used when we want to create a lambda capturing a variable
but don't want to build a trampoline because we still know its direct
signature and entry point. The trampoline will be built on demand when
the KnownProcedure is explicitly converted to a procedure cell.