Add generation of runtime predicates
This adds a concept of "writers" which are functions that take a list of
processed types and produce a map of filenames to content.
GenerateTypes will then write these to the filesystem once all the
writers complete.

This also adds a utility class called SourceBuilder which has helper
methods for generating C-like source. The only subclass right now is
CppBuilder but it's expect that a ScalaBuilder will be added that
produces properly formatted Scala.