Keep track of true alloc size in StringLikeCell
Currently we never reallocate when shrinking a string and always
reallocate when growing a string. This can lead to pathological
behaviour when repeatedly changing characters that have different UTF-8
encoding lengths than the character they're replacing.

There's only 16 bits free in StringLikeCell which is insufficient to
store the actual allocation size. Instead store the unsigned delta
between the byteCount (+ implicit NULL byte) and the actual allocation
size. This field is called allocSlackBytes.

When we allocate or reallocate a string round its size up to the next 8
byte boundary. This is intentially conservative as Scheme encourages
immutable strings and even when mutation is possible it must preserve
the character count. This means most strings will never exceed their
initial allocation size. Using 8 bytes will cause no wastage on either
Mac OS X 10.9 or glibc 2.17 on amd64.

Constant strings created by codegen and all symbols always have a slack
of 0 as they're immutable.

We reallocate a string if shrinking it would cause more than 16KiB
(MaximumAllocationSlack) of memory to be unused. This also ensures the
slack bytes count will always fit in to allocSlackBytes.