Remove UTF-8 strngs as a value type
This is completely unused by the stdlib as it uses StringCell to cache
important data about the string and allow for NULL bytes mid-string. The
current implementation is only useful for NFI but has some shortcomings.
Namely, the UTF-8 C string is not kept in sync with the boxed
representation. For example, the following would print "123":

; Set str to a UTF-8 C string containing "123"
(define str (nfi-function-returning-123-cstring))
; This will implicitly copy the UTF-8 C string to a new string cell
(string-set! str 1 #\x)
; This will implicitly copy the original string, not the modified one
(write str)

The solution is to remove UTF-8 C string support and depend on bindings
to convert back and forth. In the future explicit NFI helpers can be
provided to convert back and forth, similar to the plan with record
types.