Initial generalized support for inline cell data
This adds a new cell type called "variant" which is a a cell type that
can only be distinguished manually at runtime. The existing type-tagged
child classes have been renamed to "tagged cells" to clarify this
distinction. The intended use is to have inline and heap variants for
each cell type supporting inline data.

Initially just symbols and strings have been ported as a proof of
concept. Vectors and bytevectors are next followed by records.

This is the initial work required for issue #4