Rename "unspecific" to "unit"
The references to unspecific values in R7RS are conceptually similar to
unit values in other languages (Haskell, ML, Scala, etc). The current
name of "unspecific" is a side effect of the language in the Scheme
report and not very clear for outside developers.

Rename the type to and the value to #!unit

Closes #7