Support for LLVM IR aliases
This will be needed for generating a list of our feature identifiers.
The head of the identifier proper lists need to be aliased to a public
symbol name. We could treat the head of the list specially but in the
case of an empty list we'd need to alias anyway.