Implement (features)
This compiles in a list of feature identifiers and a simple procedure to
return them. Exporting the symbol to the stdlib as an alias was causing
weird segfaults and this method allows the feature list to be removed
from the final executable if it's not referenced.

There is no unit test because the returned lists aren't easily
searchable without (member) and related functions. These should be
implemented next.