Create GC slots on demand
Instead of creating a GC slot for every possible GC managed value only
create a slot when a value needs to be restored across a barrier. This
allows us to significantly reduce the number of slots required. We also
know the type of the value so we can remove the casts to/from %datum*.