Simple first pass of garbage collection
This provides three classes to allow the runtime to cooperate with the
GC:

- alloc::StrongRef is a strong reference to a GC managed object. These
  act as GC roots and the pointer values are automatically updated to
  their new location
- alloc::WeakRef is a weak reference to a GC managed object. These are
  not GC roots but do have their pointer values updated if the value is
  otherwise reference
- alloc::StrongRefRange will GC root a range of externally allocated
  root pointers. This has a few minor use cases related to constructing
  lists and vectors

When the current semi-space is exhausted a new semi-space is allocated
and alloc::visitCell() is used to visit each reachable cell and move it
to the new semi-space. Any weak references to live cells are then
updated to point to the value.

The runtime remains completely GC unsafe. The next commits should
resolve the most obvious issues.