Explicitly give a body for _lliby_alloc_cells
The previous method of aliasing the name didn't work (on Mac OS X 10.9
at least). There was also a potential issue on 32bit systems where the
compiler was expecting it to take std::uint64_t while the function was
actually taking a size_t

Fix by giving _lliby_alloc_cells a function body that just directly
calls allocateCells. This is a slow path so it shouldn't make any
measurable difference.