Fix GC when allocating a multi-cell range
The GC assumed every cell in the memory block was valid at GC time. This
is true when allocating a single block but will not be true when a range
allocation fails. Make sure we only update _lliby_alloc_next on success
(like the compiler does) and then use that as the real end value during
GC