Compiler support for parameterize
This required some generalization and refactor of the compiler to
support new control structures:

- Factor out most of createLambda to extractBodyDefinition. This can be
  used to process the body of any body-like definition such as (lambda),
  (parameterize), (guard), etc

- Add a ps.NestedStep that generalizes steps containing other steps.
  Most intermediate steps dealing with ps.CondBranch now deal with
  ps.NestedStep

On top of this fairly direct processing of (parameterize) is added. It's
extracted by the frontend to an et.Parameterize which is then planned in
to a ps.Parameterize. The codegen simply directly pushes a new dynamic
environment, sets the values, generates the inner steps, then pops the
state.