Implement =, <, >, <= and >=
This is implemented as a common numericCompare() template function
that's shared between each implementation. An integer and double
comparator is passed in with the compared arguments and a boolean is
returned.

This follows the R7RS recommendation of attempting to convert the values
to exact integers before falling back to IEEE floats. This means the
implementation is quite complex.