Simplify character parsing
The existing parsing put all the character handling inline with with the
main parser which caused a lot of backtracking. It also had complex
rules to satisfy R7RS that can be described more succinctly as "single
character or recognized identifier".

This makes our raw Scheme parsing 10-15% faster and our unit tests ~5%
faster.