Add datum_unchecked_cast<>
This is the same as static_assert<> in release builds but in debug
builds this will verify the actual type to ensure the cast makes sense.

This required changes to ProperList which was casting its head to
PairCell even when it's an EmptyListCell. This change in combination
with datum_unchecked_cast<> should automatically detect running past the
end of a proper list