Initial exception support
This builds on top of C++ exceptions with a special SchemeException
class. This makes it safe for exceptions to be thrown in normal runtime
code and expect destructors etc. to be called. This is in contrast with
Scheme continuations which will require great care on the C++ side to
handle arbitrary number of returns through the same code path.

A small change was required to allow copying of AbstractRefs so that
exceptions can cleanly GC root their raised object while the exception
is being handled.

This is nearly useless as it's not possible to catch an exception
without immiediately rethrowing until continuations are supported.