Add support for expecting certain output in tests
This is going to be required to test exception related things as
currently there are no way to prevent exceptions from being rethrown to
the top level. What we can do instead is exit cleanly at the last
exception handler and then examine the output produced. This is the
first step in that direction

- Add (expect-output) test type that assumes 0 or more datums will be
  printed to stdout

- Add (newline) to print a literal newline and prevent (write) from
  always emitting a newline. This is consistent with what both Racket
  and MIT Scheme do

- Add a trivial unit test for (write) and (newline) that's mostly just
  testing the (expect-output) functionality