Make adapated procedure signature take a worldptr
Adapted procedures need to have the most generic signature possible. To
make it possible to adapated procedures to wrap procedures taking a
worldptr it must take a worldptr itself.

Most of this change is merely fallout from this. (apply),
(with-exception-handler), (dynamic-wind) etc all need to have a worldptr
to properly apply Scheme procedures. This leaks in to a fair number of
other stdlib functions indirectly and serves as a good first test of
passing around worldptr