Move the strong and weak ref list to world
This had a large fallout due to the fact signalError() requires GC
rooting. This extends all the way to codegen which now needs the
worldptr for signalling errors and from there the planner which now
needs it for type casting.

At the same time perform some cleanups and refactoring on
CellRefRangeList