Pass a dummy world pointer to _lliby_alloc_cells
This is currently unused but it should represent the final changes to
the compiler required for world pointer support.

The code is a bit messy because the world pointer isn't actually used
until PlanCellAllocations which is the very last step of planning.
InferArgumentTypes and DisposeValues need to special case the world
pointer as to not optimize it out.